[spi] Rename sdi/sdo signals
SDI and SDO are good names when looking at it from a fixed role
viewpoint. Almost all MCU's however keep the same pin assignments
however when switching role, which is why the names MISO and MOSI are
used still so predominately.

E.g. if pin 1 is clk, pin 2 is mosi and pin 3 is miso, then when the
role is master, pin 2 is used to transmit data, and pin 3 to receive
data, e.g. pin 2 is sdo and pin 3 is sdi. However when the role is set
to slave, then pin 2 is now used to receive data, and pin 3 is now
transmitting data. E.g. pin 3 is now sdo and pin 2 is now sdi. This
makes things rather confusing and some switcheroo has to happen when
using SDO and SDI, MOSI and MISO however keep their naming however
exactly and do not require pins to be swapped around. The role defines
the function.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>